cleaned up utils folder a little bit
