+++ swyft/saveable.py
