+++ swyft/store/dataset.py
@@ -6,9 +6,9 @@ import numpy as np
 import torch
 
 import swyft
+from swyft.saveable import StateDictSaveable
 from swyft.types import ObsType, ParameterNamesType, PathType
 from swyft.utils.array import array_to_tensor
-from swyft.utils.saveable import StateDictSaveable
 
 log = logging.getLogger(__name__)
