+++ swyft/utils/__init__.py
@@ -1,10 +1,8 @@
 from swyft.utils.array import *
-from swyft.utils.device import *
-from swyft.utils.mutils import *
-from swyft.utils.parameters import *
+from swyft.utils.marginalutils import *
+from swyft.utils.mass import *
 from swyft.utils.plot import *
-from swyft.utils.plot import plot_1d, plot_corner, plot_empirical_mass
 from swyft.utils.utils import *
-from swyft.utils.wmutils import *
+from swyft.utils.weightedmarginalutils import *
 
 __all__ = ["plot_1d", "plot_corner", "plot_empirical_mass"]
